test: 로또 테스트 작성
